Reworking imports to be more compatible.  Renaming query to be more descriptive.
