+++ dictorm/dictorm.py
@@ -10,14 +10,14 @@ from json import dumps
 from copy import deepcopy
 
 try: # pragma: no cover
-    from dictorm.query import Select, Insert, Update, Delete, And
-    from dictorm.query import Column, Expression, Logical
+    from dictorm.pg import Select, Insert, Update, Delete, And
+    from dictorm.pg import Column, Expression, Logical
     from dictorm.sqlite import Insert as SqliteInsert
     from dictorm.sqlite import Column as SqliteColumn
     from dictorm.sqlite import Update as SqliteUpdate
 except ImportError: # pragma: no cover
-    from .query import Select, Insert, Update, Delete, And
-    from .query import Column, Expression, Logical
+    from .pg import Select, Insert, Update, Delete, And
+    from .pg import Column, Expression, Logical
     from .sqlite import Insert as SqliteInsert
     from .sqlite import Column as SqliteColumn
     from .sqlite import Update as SqliteUpdate
